Wait until the first pageshow to attach instrumentation.
We found that accessing certain elements of the DOM before the XUL
overlay is attached can prevent the overlay from being attached.
Strangely, merely indexing one of the base elements (as returned from
document.getElementsByTagName("*")) was sufficient to prevent the
overlay from attaching. This problem manifested by the groupbox elements
on the bridge settings and proxy settings pages not displaying.

We could access almost every element, but touching e.g. element #45,
which was the groupbox with id bridgeSpecificSettings, would prevent the
overlay from attaching.