Attach instrumentation to the root document element.
This is better than walking the tree and attaching to individual
elements. The event bubbles up to the root element but gets logged with
the element that was (for example) clicked on.

This also seems to sidestep the problem described in d6d2e2d7e3, where
indexing certin elements that were to have overlays attached would cause
the overlays to fail.

This additionally fixes a problem with the workaround in d6d2e2d7e3,
which is that the first pageshow would not be logged, because the
instrumentation was being added in the pageshow callback of the first
wizardpage.