Don't hide "next", "extra1", "extra2" on the progress screen.
The main reason for this change is because if bootstrapping failed, and
you clicked "Back" to the bridge screen, there would be no "Next"
button.

"next" doesn't need to be hidden because the progress page doesn't have
a "next" attribute in the XUL, so the next button is automatically
hidden on that page.

"extra1" isn't used in the code.

"extra2" is the "Copy Tor Log to Clipboard" button, which we want to
keep visible if it is visible already.